+++ man-roxygen/opt-slide-params.R
@@ -10,8 +10,5 @@
 #'   to provide output column names; if you want to customize the output column
 #'   names, use [`dplyr::rename`] after the slide.
 #' @param as_list_col Not supported. Included to match `epi_slide` interface.
-#' @param new_col_name Character vector indicating the name(s) of the new
-#'  column(s) that will contain the derivative values. Default
-#'  is "slide_value"; note that setting `new_col_name` equal to any existing
-#'  column names will overwrite those columns. If `names_sep` is `NULL`,
-#'  `new_col_name` must be the same length as `col_names`.
+#' @param new_col_name Not supported. Included to match `epi_slide` interface.
+#' @param names_sep Not supported. Included to match `epi_slide` interface.