+++ tests/testthat/test-epi_slide.R
@@ -52,19 +52,19 @@ basic_mean_result <- tibble::tribble(
 test_that("`before` and `after` are both vectors of length 1", {
   expect_error(
     epi_slide(grouped, f, before = c(0, 1), after = 0, ref_time_values = test_date + 3),
-    "Expected `before` to be a scalar value."
+    "Slide function expected `before` to be a scalar value."
   )
   expect_error(
     epi_slide(grouped, f, before = 1, after = c(0, 1), ref_time_values = test_date + 3),
-    "Expected `after` to be a scalar value."
+    "Slide function expected `after` to be a scalar value."
   )
   expect_error(
     epi_slide_mean(grouped, col_names = value, before = c(0, 1), after = 0, ref_time_values = test_date + 3),
-    "Expected `before` to be a scalar value."
+    "Slide function expected `before` to be a scalar value."
   )
   expect_error(
     epi_slide_mean(grouped, col_names = value, before = 1, after = c(0, 1), ref_time_values = test_date + 3),
-    "Expected `after` to be a scalar value."
+    "Slide function expected `after` to be a scalar value."
   )
 })
 
@@ -107,64 +107,64 @@ test_that("Test errors/warnings for discouraged features", {
 test_that("Both `before` and `after` must be non-NA, non-negative, integer-compatible", {
   expect_error(
     epi_slide(grouped, f, before = -1L, ref_time_values = test_date + 2L),
-    "Expected `before` to be a difftime with units in days or a non-negative integer."
+    "Slide function expected `before` to be a difftime with units in days or non-negative integer or Inf."
   )
   expect_error(
     epi_slide(grouped, f, after = -1L, ref_time_values = test_date + 2L),
-    "Expected `after` to be a difftime with units in days or a non-negative integer."
+    "Slide function expected `after` to be a difftime with units in days or non-negative integer or Inf."
   )
   expect_error(epi_slide(grouped, f, before = "a", after = days_dt, ref_time_values = test_date + 2L),
-    regexp = "Expected `before` to be a difftime with units in days or a non-negative integer."
+    regexp = "Slide function expected `before` to be a difftime with units in days or non-negative integer or Inf."
   )
   expect_error(epi_slide(grouped, f, before = days_dt, after = "a", ref_time_values = test_date + 2L),
-    regexp = "Expected `after` to be a difftime with units in days or a non-negative integer."
+    regexp = "Slide function expected `after` to be a difftime with units in days or non-negative integer or Inf."
   )
   expect_error(epi_slide(grouped, f, before = 0.5, after = days_dt, ref_time_values = test_date + 2L),
-    regexp = "Expected `before` to be a difftime with units in days or a non-negative integer."
+    regexp = "Slide function expected `before` to be a difftime with units in days or non-negative integer or Inf."
   )
   expect_error(epi_slide(grouped, f, before = days_dt, after = 0.5, ref_time_values = test_date + 2L),
-    regexp = "Expected `after` to be a difftime with units in days or a non-negative integer."
+    regexp = "Slide function expected `after` to be a difftime with units in days or non-negative integer or Inf."
   )
   expect_error(
     epi_slide(grouped, f, before = NA, after = 1L, ref_time_values = test_date + 2L),
-    "Expected `before` to be a scalar value."
+    "Slide function expected `before` to be a scalar value."
   )
   expect_error(
     epi_slide(grouped, f, before = days_dt, after = NA, ref_time_values = test_date + 2L),
-    "Expected `after` to be a scalar value."
+    "Slide function expected `after` to be a scalar value."
   )
 
   expect_error(
     epi_slide_mean(grouped, col_names = value, before = -1L, ref_time_values = test_date + 2L),
-    "Expected `before` to be a difftime with units in days or a non-negative integer."
+    "Slide function expected `before` to be a difftime with units in days or non-negative integer or Inf."
   )
   expect_error(
     epi_slide_mean(grouped, col_names = value, after = -1L, ref_time_values = test_date + 2L),
-    "Expected `after` to be a difftime with units in days or a non-negative integer."
+    "Slide function expected `after` to be a difftime with units in days or non-negative integer or Inf."
   )
   expect_error(
     epi_slide_mean(grouped, col_names = value, before = "a", ref_time_values = test_date + 2L),
-    regexp = "Expected `before` to be a difftime with units in days or a non-negative integer."
+    regexp = "Slide function expected `before` to be a difftime with units in days or non-negative integer or Inf."
   )
   expect_error(
     epi_slide_mean(grouped, col_names = value, after = "a", ref_time_values = test_date + 2L),
-    regexp = "Expected `after` to be a difftime with units in days or a non-negative integer."
+    regexp = "Slide function expected `after` to be a difftime with units in days or non-negative integer or Inf."
   )
   expect_error(
     epi_slide_mean(grouped, col_names = value, before = 0.5, ref_time_values = test_date + 2L),
-    regexp = "Expected `before` to be a difftime with units in days or a non-negative integer."
+    regexp = "Slide function expected `before` to be a difftime with units in days or non-negative integer or Inf."
   )
   expect_error(
     epi_slide_mean(grouped, col_names = value, after = 0.5, ref_time_values = test_date + 2L),
-    regexp = "Expected `after` to be a difftime with units in days or a non-negative integer."
+    regexp = "Slide function expected `after` to be a difftime with units in days or non-negative integer or Inf."
   )
   expect_error(
     epi_slide_mean(grouped, col_names = value, before = NA, after = days_dt, ref_time_values = test_date + 2L),
-    "Expected `before` to be a scalar value."
+    "Slide function expected `before` to be a scalar value."
   )
   expect_error(
     epi_slide_mean(grouped, col_names = value, before = days_dt, after = NA, ref_time_values = test_date + 2L),
-    "Expected `after` to be a scalar value."
+    "Slide function expected `after` to be a scalar value."
   )
 
   # Non-integer-class but integer-compatible values are allowed:
@@ -434,7 +434,7 @@ test_that("`ref_time_values` + `all_rows = TRUE` works", {
     ),
     basic_full_result %>%
       dplyr::mutate(slide_value = dplyr::if_else(time_value %in% (test_date + c(2L, 8L)),
-        slide_value, NA_integer_
+        slide_value, NA_real_ # (`^` outputs numeric)
       ))
   )
 
@@ -444,7 +444,7 @@ test_that("`ref_time_values` + `all_rows = TRUE` works", {
     ) %>%
       epi_slide_mean(
         value,
-        before = 6 * days_dt, names_sep = NULL, na.rm = TRUE
+        before = 6 * days_dt, na.rm = TRUE
       ),
     basic_mean_result %>%
       rename(slide_value_value = slide_value)
@@ -472,7 +472,7 @@ test_that("`ref_time_values` + `all_rows = TRUE` works", {
       ),
     basic_mean_result %>%
       dplyr::mutate(slide_value_value = dplyr::if_else(time_value %in% (test_date + c(2L, 8L)),
-        slide_value, NA_integer_
+        slide_value, NA_real_
       )) %>%
       select(-slide_value)
   )
@@ -498,7 +498,7 @@ test_that("`ref_time_values` + `all_rows = TRUE` works", {
     ),
     basic_full_result %>%
       dplyr::mutate(slide_value = dplyr::if_else(time_value %in% (test_date + c(2L, 8L)),
-        slide_value, NA_integer_
+        slide_value, NA_real_
       )) %>%
       dplyr::rename(slide_value_value = slide_value)
   )
@@ -584,7 +584,7 @@ test_that("`ref_time_values` + `all_rows = TRUE` works", {
       unnest(slide_value, names_sep = "_"),
     basic_full_result %>%
       dplyr::mutate(slide_value = dplyr::if_else(time_value %in% (test_date + c(2L, 8L)),
-        slide_value, NA_integer_
+        slide_value, NA_real_
       )) %>%
       dplyr::rename(slide_value_value = slide_value)
   )
@@ -1274,7 +1274,6 @@ test_that("`epi_slide_opt` errors when passed non-`data.table`, non-`slider` fun
   )
 })
 
-
 multi_columns <- dplyr::bind_rows(
   dplyr::tibble(geo_value = "ak", time_value = test_date + 1:200, value = 1:200, value2 = -1:-200),
   dplyr::tibble(geo_value = "al", time_value = test_date + 1:5, value = -(1:5), value2 = 1:5)
@@ -1298,4 +1297,109 @@ test_that("no dplyr warnings from selecting multiple columns", {
     multi_slid_select <- epi_slide_mean(multi_columns, starts_with("value"), before = 3L)
   )
   expect_equal(multi_slid_select, multi_slid)
+})
+  
+test_that("Inf works in before/after in slide and slide_opt", {
+  # Daily data
+  df <- dplyr::bind_rows(
+    dplyr::tibble(geo_value = "ak", time_value = test_date + 1:200, value = 1:200),
+    dplyr::tibble(geo_value = "al", time_value = test_date + 1:5, value = -(1:5))
+  ) %>%
+    as_epi_df()
+  expect_equal(
+    df %>%
+      group_by(geo_value) %>%
+      epi_slide(
+        before = Inf,
+        slide_value = sum(value)
+      ),
+    df %>%
+      group_by(geo_value) %>%
+      epi_slide(
+        before = 365000,
+        slide_value = sum(value)
+      )
+  )
+  expect_equal(
+    df %>%
+      group_by(geo_value) %>%
+      epi_slide_opt(
+        before = Inf,
+        f = data.table::frollsum,
+        col_names = value
+      ),
+    df %>%
+      group_by(geo_value) %>%
+      epi_slide(
+        before = 365000,
+        slide_value_value = sum(value)
+      )
+  )
+  expect_equal(
+    df %>%
+      group_by(geo_value) %>%
+      epi_slide_opt(
+        before = Inf,
+        f = slider::slide_sum,
+        col_names = value
+      ),
+    df %>%
+      group_by(geo_value) %>%
+      epi_slide(
+        before = 365000,
+        slide_value_value = sum(value)
+      )
+  )
+
+  # Weekly data
+  df <- dplyr::bind_rows(
+    dplyr::tibble(geo_value = "ak", time_value = test_date + 1:200 * 7, value = 1:200),
+    dplyr::tibble(geo_value = "al", time_value = test_date + 1:5 * 7, value = -(1:5))
+  ) %>%
+    as_epi_df()
+
+  expect_equal(
+    df %>%
+      group_by(geo_value) %>%
+      epi_slide(
+        before = Inf,
+        slide_value = sum(value)
+      ),
+    df %>%
+      group_by(geo_value) %>%
+      epi_slide(
+        before = 365000 * weeks_dt,
+        slide_value = sum(value)
+      )
+  )
+  expect_equal(
+    df %>%
+      group_by(geo_value) %>%
+      epi_slide_opt(
+        col_names = value,
+        f = data.table::frollsum,
+        before = Inf
+      ),
+    df %>%
+      group_by(geo_value) %>%
+      epi_slide(
+        before = 365000 * weeks_dt,
+        slide_value_value = sum(value)
+      )
+  )
+  expect_equal(
+    df %>%
+      group_by(geo_value) %>%
+      epi_slide_opt(
+        before = Inf,
+        f = slider::slide_sum,
+        col_names = value
+      ),
+    df %>%
+      group_by(geo_value) %>%
+      epi_slide(
+        before = 365000 * weeks_dt,
+        slide_value_value = sum(value)
+      )
+  )
 })