+++ tests/testthat/test-methods-epi_df.R
@@ -173,7 +173,142 @@ test_that("Renaming columns while grouped gives appropriate colnames and metadat
   expect_identical(attr(renamed_gedf1, "metadata")$other_keys, c("age_group"))
   # renaming using select
   renamed_gedf2 <- gedf %>%
-    as_epi_df(additional_metadata = list(other_keys = "age")) %>%
     select(geo_value, time_value, age_group = age, value)
   expect_identical(renamed_gedf1, renamed_gedf2)
+})
+
+test_that("Additional `select` on `epi_df` tests", {
+  edf <- tibble::tibble(geo_value = "ak", time_value = as.Date("2020-01-01"), age = 1, value = 1) %>%
+    as_epi_df(additional_metadata = list(other_keys = "age"))
+
+  # Dropping a non-geo_value epikey column doesn't decay, though maybe it
+  # should, since you'd expect that to possibly result in multiple rows per
+  # epikey (though not in this toy case), and while we don't require that, we
+  # sort of expect it:
+  edf_not_decayed <- edf %>%
+    select(geo_value, time_value, value)
+  expect_class(edf_not_decayed, "epi_df")
+  expect_identical(attr(edf_not_decayed, "metadata")$other_keys, character(0L))
+
+  # Dropping geo_value does decay:
+  edf_decayed <- edf %>%
+    select(age, time_value, value)
+  expect_false(inherits(edf_decayed, "epi_df"))
+  expect_identical(attr(edf_decayed, "metadata"), NULL)
+})
+
+test_that("complete.epi_df works", {
+  start_date <- as.Date("2020-01-01")
+  daily_edf <- tibble::tribble(
+    ~geo_value, ~time_value, ~value,
+    1, start_date + 1, 1,
+    1, start_date + 3, 3,
+    2, start_date + 2, 2,
+    2, start_date + 3, 3,
+  ) %>%
+    as_epi_df(as_of = start_date + 3)
+  # Complete without grouping puts all the geo_values on the same min and max
+  # time_value index
+  expect_identical(
+    daily_edf %>%
+      complete(geo_value, time_value = full_seq(time_value, period = 1)),
+    tibble::tribble(
+      ~geo_value, ~time_value, ~value,
+      1, start_date + 1, 1,
+      1, start_date + 2, NA,
+      1, start_date + 3, 3,
+      2, start_date + 1, NA,
+      2, start_date + 2, 2,
+      2, start_date + 3, 3,
+    ) %>%
+      as_epi_df(as_of = start_date + 3)
+  )
+  # Complete with grouping puts all the geo_values on individual min and max
+  # time_value indices
+  expect_identical(
+    daily_edf %>%
+      group_by(geo_value) %>%
+      complete(time_value = full_seq(time_value, period = 1)),
+    tibble::tribble(
+      ~geo_value, ~time_value, ~value,
+      1, start_date + 1, 1,
+      1, start_date + 2, NA,
+      1, start_date + 3, 3,
+      2, start_date + 2, 2,
+      2, start_date + 3, 3,
+    ) %>%
+      as_epi_df(as_of = start_date + 3) %>%
+      group_by(geo_value)
+  )
+  # Complete has explicit=TRUE by default, but if it's FALSE, then complete only fills the implicit gaps
+  # not those that are explicitly NA
+  daily_edf <- tibble::tribble(
+    ~geo_value, ~time_value, ~value,
+    1, start_date + 1, 1,
+    1, start_date + 2, NA,
+    1, start_date + 3, 3,
+    2, start_date + 2, 2,
+    2, start_date + 3, 3,
+  ) %>%
+    as_epi_df(as_of = start_date + 3)
+  expect_identical(
+    daily_edf %>%
+      complete(geo_value, time_value = full_seq(time_value, period = 1), fill = list(value = 0), explicit = FALSE),
+    tibble::tribble(
+      ~geo_value, ~time_value, ~value,
+      1, start_date + 1, 1,
+      1, start_date + 2, NA,
+      1, start_date + 3, 3,
+      2, start_date + 1, 0,
+      2, start_date + 2, 2,
+      2, start_date + 3, 3,
+    ) %>%
+      as_epi_df(as_of = start_date + 3)
+  )
+  # Complete works for weekly data and can take a fill value
+  # No grouping
+  weekly_edf <- tibble::tribble(
+    ~geo_value, ~time_value, ~value,
+    1, start_date + 1, 1,
+    1, start_date + 15, 3,
+    2, start_date + 8, 2,
+    2, start_date + 15, 3,
+  ) %>%
+    as_epi_df(as_of = start_date + 3)
+  expect_identical(
+    weekly_edf %>%
+      complete(geo_value,
+        time_value = full_seq(time_value, period = 7),
+        fill = list(value = 0)
+      ),
+    tibble::tribble(
+      ~geo_value, ~time_value, ~value,
+      1, start_date + 1, 1,
+      1, start_date + 8, 0,
+      1, start_date + 15, 3,
+      2, start_date + 1, 0,
+      2, start_date + 8, 2,
+      2, start_date + 15, 3,
+    ) %>%
+      as_epi_df(as_of = start_date + 3)
+  )
+  # With grouping
+  expect_identical(
+    weekly_edf %>%
+      group_by(geo_value) %>%
+      complete(
+        time_value = full_seq(time_value, period = 7),
+        fill = list(value = 0)
+      ),
+    tibble::tribble(
+      ~geo_value, ~time_value, ~value,
+      1, start_date + 1, 1,
+      1, start_date + 8, 0,
+      1, start_date + 15, 3,
+      2, start_date + 8, 2,
+      2, start_date + 15, 3,
+    ) %>%
+      as_epi_df(as_of = start_date + 3) %>%
+      group_by(geo_value)
+  )
 })